third_party/blink/web_tests/external/wpt/resource-timing/initiator-type/script-expected.txt

This is a testharness.js-based test.
[FAIL] The initiator type for script imported from another script must be 'script'
  assert_equals: expected "script" but got "other"
Harness: the test ran to completion.
